Read the metadata of a Parquet file

This function should work on all files, even if read_parquet() is unabled to read them, because of an unsupported schema, encoding, compression or other reason.

read_parquet_metadata(file)

Arguments

file

Path to a Parquet file.

Value

A named list with entries:

file_meta_data: a named list with file meta data:

version: Parquet version, an integer.

num_rows: total number of rows.

key_value_metadata: a data frame with two character columns called key and value. This is the key-value metadata of the file. Arrow stores its schema here.

created_by: A string scalar, usually the name of the software that created the file.

encryption_algorithm: not implemented yet.

footer_signing_key_metadata: not implemented yet.

schema: data frame, the schema of the file. It has one row for each node (inner node or leaf node). For flat files this means one root node (inner node), always the first one, and then one row for each "real" column. For nested schemas, the rows are in depth-first search order. Most important columns are:

name: column name.

type: data type. One of the low level data types.

type_length: length for fixed length byte arrays.

repettion_type: character, one of REQUIRED, OPTIONAL or REPEATED.

logical_type: a list column, the logical types of the columns. An element has at least an entry called type, and potentially additional entries, e.g. bit_width, is_signed, etc.

num_children: number of child nodes. Should be a non-negative integer for the root node, and NA for a leaf node.

$row_groups: a data frame, information about the row groups.

$column_chunks: a data frame, information about all column chunks.

See also

Examples

file_name <- system.file("extdata/userdata1.parquet", package = "miniparquet")
miniparquet::read_parquet_metadata(file_name)
#> $file_meta_data
#> $file_meta_data$version
#> [1] 1
#> 
#> $file_meta_data$num_rows
#> [1] 1000
#> 
#> $file_meta_data$key_value_metadata
#> # A data frame: 0 × 2
#> # ℹ 2 variables: key <chr>, value <chr>
#> 
#> $file_meta_data$created_by
#> [1] "parquet-mr version 1.8.1 (build 4aba4dae7bb0d4edbcf7923ae1339f28fd3f7fcf)"
#> 
#> $file_meta_data$encryption_algorithm
#> NULL
#> 
#> $file_meta_data$footer_signing_key_metadata
#> NULL
#> 
#> 
#> $schema
#> # A data frame: 14 × 10
#>    name            type  type_length repetition_type converted_type logical_type
#>    <chr>           <chr>       <int> <chr>           <chr>          <I<list>>   
#>  1 hive_schema     NA             NA NA              NA             <NULL>      
#>  2 registration_d… INT96          NA OPTIONAL        NA             <NULL>      
#>  3 id              INT32          NA OPTIONAL        NA             <NULL>      
#>  4 first_name      BYTE…          NA OPTIONAL        UTF8           <NULL>      
#>  5 last_name       BYTE…          NA OPTIONAL        UTF8           <NULL>      
#>  6 email           BYTE…          NA OPTIONAL        UTF8           <NULL>      
#>  7 gender          BYTE…          NA OPTIONAL        UTF8           <NULL>      
#>  8 ip_address      BYTE…          NA OPTIONAL        UTF8           <NULL>      
#>  9 cc              BYTE…          NA OPTIONAL        UTF8           <NULL>      
#> 10 country         BYTE…          NA OPTIONAL        UTF8           <NULL>      
#> 11 birthdate       BYTE…          NA OPTIONAL        UTF8           <NULL>      
#> 12 salary          DOUB…          NA OPTIONAL        NA             <NULL>      
#> 13 title           BYTE…          NA OPTIONAL        UTF8           <NULL>      
#> 14 comments        BYTE…          NA OPTIONAL        UTF8           <NULL>      
#> # ℹ 4 more variables: num_children <int>, scale <int>, precision <int>,
#> #   field_id <int>
#> 
#> $row_groups
#> # A data frame: 1 × 6
#>      id total_byte_size num_rows file_offset total_compressed_size ordinal
#>   <int>           <dbl>    <dbl>       <dbl>                 <dbl>   <int>
#> 1     0          112492     1000          NA                    NA      NA
#> 
#> $column_chunks
#> # A data frame: 13 × 17
#>    row_group file_path file_offset offset_index_offset offset_index_length
#>        <int> <chr>           <dbl>               <dbl>               <int>
#>  1         0 NA                  4                  NA                  NA
#>  2         0 NA              13274                  NA                  NA
#>  3         0 NA              17317                  NA                  NA
#>  4         0 NA              20305                  NA                  NA
#>  5         0 NA              23855                  NA                  NA
#>  6         0 NA              48543                  NA                  NA
#>  7         0 NA              48871                  NA                  NA
#>  8         0 NA              66223                  NA                  NA
#>  9         0 NA              81686                  NA                  NA
#> 10         0 NA              84136                  NA                  NA
#> 11         0 NA              95403                  NA                  NA
#> 12         0 NA             103034                  NA                  NA
#> 13         0 NA             108208                  NA                  NA
#> # ℹ 12 more variables: column_index_offset <dbl>, column_index_length <int>,
#> #   type <int>, encodings <I<list>>, path_in_schema <I<list>>, codec <chr>,
#> #   num_values <dbl>, total_uncompressed_size <dbl>,
#> #   total_compressed_size <dbl>, data_page_offset <dbl>,
#> #   index_page_offset <dbl>, dictionary_page_offset <dbl>
#>