Read the metadata of a Parquet file

This function should work on all files, even if read_parquet() is unabled to read them, because of an unsupported schema, encoding, compression or other reason.

read_parquet_metadata(file)

Arguments

file

Path to a Parquet file.

Value

A named list with entries:

file_meta_data: a data frame with file meta data:

file_name: file name.

version: Parquet version, an integer.

num_rows: total number of rows.

key_value_metadata: list column of a data frames with two character columns called key and value. This is the key-value metadata of the file. Arrow stores its schema here.

created_by: A string scalar, usually the name of the software that created the file.

schema: data frame, the schema of the file. It has one row for each node (inner node or leaf node). For flat files this means one root node (inner node), always the first one, and then one row for each "real" column. For nested schemas, the rows are in depth-first search order. Most important columns are:

file_name: file name.

name: column name.

type: data type. One of the low level data types.

type_length: length for fixed length byte arrays.

repettion_type: character, one of REQUIRED, OPTIONAL or REPEATED.

logical_type: a list column, the logical types of the columns. An element has at least an entry called type, and potentially additional entries, e.g. bit_width, is_signed, etc.

num_children: number of child nodes. Should be a non-negative integer for the root node, and NA for a leaf node.

$row_groups: a data frame, information about the row groups.

$column_chunks: a data frame, information about all column chunks, across all row groups. Some important columns:

file_name: file name.

row_group: which row group this chunk belongs to.

column: which leaf column this chunks belongs to. The order is the same as in $schema, but only leaf columns (i.e. columns with NA children) are counted.

file_path: which file the chunk is stored in. NA means the same file.

file_offset: where the column chunk begins in the file.

type: low level parquet data type.

encodings: encodings used to store this chunk. It is a list column of character vectors of encoding names. Current possible encodings: "PLAIN", "GROUP_VAR_INT", "PLAIN_DICTIONARY", "RLE", "BIT_PACKED", "DELTA_BINARY_PACKED", "DELTA_LENGTH_BYTE_ARRAY", "DELTA_BYTE_ARRAY", "RLE_DICTIONARY".

path_in_scema: list column of character vectors. It is simply the path from the root node. It is simply the column name for flat schemas.

codec: compression codec used for the column chunk. Possible values are: "UNCOMPRESSED", "SNAPPY", "GZIP", "LZO", "BROTLI", "LZ4", "ZSTD".

num_values: number of values in this column chunk.

total_uncompressed_size: total uncompressed size in bytes.

total_compressed_size: total compressed size in bytes.

data_page_offset: absolute position of the first data page of the column chunk in the file.

index_page_offset: absolute position of the first index page of the column chunk in the file, or NA if there are no index pages.

dictionary_page_offset: absolute position of the first dictionary page of the column chunk in the file, or NA if there are no dictionary pages.

See also

Examples

file_name <- system.file("extdata/userdata1.parquet", package = "miniparquet")
miniparquet::read_parquet_metadata(file_name)
#> $file_meta_data
#> # A data frame: 1 × 5
#>   file_name                       version num_rows key_value_metadata created_by
#>   <chr>                             <int>    <dbl> <I<list>>          <chr>     
#> 1 /home/runner/work/_temp/Librar…       1     1000 <tbl [0 × 2]>      parquet-m…
#> 
#> $schema
#> # A data frame: 14 × 11
#>    file_name name  type  type_length repetition_type converted_type logical_type
#>    <chr>     <chr> <chr>       <int> <chr>           <chr>          <I<list>>   
#>  1 /home/ru… hive… NA             NA NA              NA             <NULL>      
#>  2 /home/ru… regi… INT96          NA OPTIONAL        NA             <NULL>      
#>  3 /home/ru… id    INT32          NA OPTIONAL        NA             <NULL>      
#>  4 /home/ru… firs… BYTE…          NA OPTIONAL        UTF8           <NULL>      
#>  5 /home/ru… last… BYTE…          NA OPTIONAL        UTF8           <NULL>      
#>  6 /home/ru… email BYTE…          NA OPTIONAL        UTF8           <NULL>      
#>  7 /home/ru… gend… BYTE…          NA OPTIONAL        UTF8           <NULL>      
#>  8 /home/ru… ip_a… BYTE…          NA OPTIONAL        UTF8           <NULL>      
#>  9 /home/ru… cc    BYTE…          NA OPTIONAL        UTF8           <NULL>      
#> 10 /home/ru… coun… BYTE…          NA OPTIONAL        UTF8           <NULL>      
#> 11 /home/ru… birt… BYTE…          NA OPTIONAL        UTF8           <NULL>      
#> 12 /home/ru… sala… DOUB…          NA OPTIONAL        NA             <NULL>      
#> 13 /home/ru… title BYTE…          NA OPTIONAL        UTF8           <NULL>      
#> 14 /home/ru… comm… BYTE…          NA OPTIONAL        UTF8           <NULL>      
#> # ℹ 4 more variables: num_children <int>, scale <int>, precision <int>,
#> #   field_id <int>
#> 
#> $row_groups
#> # A data frame: 1 × 7
#>   file_name        id total_byte_size num_rows file_offset total_compressed_size
#>   <chr>         <int>           <dbl>    <dbl>       <dbl>                 <dbl>
#> 1 /home/runner…     0          112492     1000          NA                    NA
#> # ℹ 1 more variable: ordinal <int>
#> 
#> $column_chunks
#> # A data frame: 13 × 19
#>    file_name          row_group column file_path file_offset offset_index_offset
#>    <chr>                  <int>  <int> <chr>           <dbl>               <dbl>
#>  1 /home/runner/work…         0      0 NA                  4                  NA
#>  2 /home/runner/work…         0      1 NA              13274                  NA
#>  3 /home/runner/work…         0      2 NA              17317                  NA
#>  4 /home/runner/work…         0      3 NA              20305                  NA
#>  5 /home/runner/work…         0      4 NA              23855                  NA
#>  6 /home/runner/work…         0      5 NA              48543                  NA
#>  7 /home/runner/work…         0      6 NA              48871                  NA
#>  8 /home/runner/work…         0      7 NA              66223                  NA
#>  9 /home/runner/work…         0      8 NA              81686                  NA
#> 10 /home/runner/work…         0      9 NA              84136                  NA
#> 11 /home/runner/work…         0     10 NA              95403                  NA
#> 12 /home/runner/work…         0     11 NA             103034                  NA
#> 13 /home/runner/work…         0     12 NA             108208                  NA
#> # ℹ 13 more variables: offset_index_length <int>, column_index_offset <dbl>,
#> #   column_index_length <int>, type <chr>, encodings <I<list>>,
#> #   path_in_schema <I<list>>, codec <chr>, num_values <dbl>,
#> #   total_uncompressed_size <dbl>, total_compressed_size <dbl>,
#> #   data_page_offset <dbl>, index_page_offset <dbl>,
#> #   dictionary_page_offset <dbl>
#>